add format script
