+++ packages/typespec-python/scripts/format.ts
@@ -1,0 +1,5 @@
+/* eslint-disable no-console */
+
+import { runCommand } from "./utils.js";
+
+runCommand("black .", "black");
