+++ packages/typespec-python/scripts/utils.ts
@@ -1,0 +1,26 @@
+/** */
+import { exec } from "child_process";
+import { existsSync } from "fs";
+import { dirname, join } from "path";
+import { fileURLToPath } from "url";
+import chalk from "chalk";
+
+// Function to run a command and log the output
+export function runCommand(command: string, prettyName: string) {
+    let pythonPath = join(dirname(fileURLToPath(import.meta.url)), "..", "venv/");
+    if (existsSync(join(pythonPath, "bin"))) {
+        pythonPath = join(pythonPath, "bin", "python");
+    } else if (existsSync(join(pythonPath, "Scripts"))) {
+        pythonPath = join(pythonPath, "Scripts", "python");
+    } else {
+        throw new Error(pythonPath);
+    }
+    command = `${pythonPath} -m ${command}`;
+    exec(command, (error, stdout, stderr) => {
+        if (error || stderr) {
+            console.error(chalk.red(`Error executing ${command}: ${stdout}`));
+            return;
+        }
+        console.log(chalk.green(`${prettyName} passed`));
+    });
+}
